[giac] visualc++ warning cleanup bug fix for matrix_norm(.,2)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51553 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ giac/src/giac/cpp/global.cc
@@ -4375,7 +4375,7 @@ unsigned int ConvertUTF8toUTF16 (
   // count number of bytes required to save g in a file
   static size_t countfunction(void const* p, size_t nbBytes,size_t NbElements, void *file)
   {
-    (*(unsigned *)file)+= nbBytes*NbElements;
+    (*(unsigned *)file)+= unsigned(nbBytes*NbElements);
     return nbBytes*NbElements;
   }
   unsigned archive_count(const gen & g,GIAC_CONTEXT){
